vision_hwr_agent/requirements.txt
# /composable-orchestrator/vision_hwr_agent/requirements.txt

pika
pillow
pytesseract